Move members dataminimisation to apps
