+++ website/members/apps.py
@@ -1,7 +1,13 @@
 from django.apps import AppConfig
+from django.db.models import Count, Exists, OuterRef, Q
 from django.template.loader import render_to_string
 from django.urls import reverse
+from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
+
+from registrations.models import Renewal
+
+from .models.member import Member
 
 
 class MembersConfig(AppConfig):
@@ -90,3 +96,59 @@ class MembersConfig(AppConfig):
                 }
             )
         return announcements
+
+    def execute_data_minimisation(self, dry_run=False, members=None) -> list[Member]:
+        """Clean the profiles of members/users of whom the last membership ended at least 90 days ago.
+
+        :param dry_run: does not really remove data if True
+        :param members: queryset of members to process, optional
+        :return: list of processed members
+        """
+        if not members:
+            members = Member.objects
+        members = (
+            members.annotate(membership_count=Count("membership"))
+            .exclude(
+                (
+                    Q(membership__until__isnull=True)
+                    | Q(membership__until__gt=timezone.now().date())
+                )
+                & Q(membership_count__gt=0)
+            )
+            .exclude(
+                Exists(
+                    Renewal.objects.filter(member__id=OuterRef("pk")).exclude(
+                        status__in=(
+                            Renewal.STATUS_ACCEPTED,
+                            Renewal.STATUS_REJECTED,
+                        )
+                    )
+                )
+            )
+            .distinct()
+            .prefetch_related("membership_set", "profile")
+        )
+        deletion_period = timezone.now().date() - timezone.timedelta(days=90)
+        processed_members = []
+        for member in members:
+            if (
+                member.latest_membership is None
+                or member.latest_membership.until <= deletion_period
+            ):
+                processed_members.append(member)
+                profile = member.profile
+                profile.student_number = None
+                profile.phone_number = None
+                profile.address_street = None
+                profile.address_street2 = None
+                profile.address_postal_code = None
+                profile.address_city = None
+                profile.address_country = None
+                profile.birthday = None
+                profile.emergency_contact_phone_number = None
+                profile.emergency_contact = None
+                profile.is_minimized = True
+                if not dry_run:
+                    profile.save()
+
+        return processed_members
